+++ src/components/AdminSettings.vue
@@ -27,19 +27,12 @@
 				:is-complete="isAuthorizationSettingFormComplete"
 				:is-disabled="isAuthorizationSettingFormInDisabledMode"
 				:is-dark-theme="isDarkTheme"
-				:has-error="!isOIDCAppInstalledAndEnabled || !isOIDCAppSupported" />
+				:has-error="!hasEnabledSupportedUserOidcApp" />
 			<ErrorNote
-				v-if="!isOIDCAppInstalledAndEnabled"
-				:error-title="messagesFmt.appNotInstalled('user_oidc')"
-				:error-message="messages.appRequiredForOIDCMethod"
+				v-if="!hasEnabledSupportedUserOidcApp"
+				:error-title="messagesFmt.appNotEnabledOrUnsupported('user_oidc', getMinSupportedUserOidcVersion)"
 				:error-link="appLinks.user_oidc.installLink"
-				:error-link-label="messages.downloadAndEnableApp" />
-			<ErrorNote
-				v-if="isOIDCAppInstalledAndEnabled && !isOIDCAppSupported"
-				:error-title="messagesFmt.appNotSupported('user_oidc')"
-				:error-message="messagesFmt.minimumVersionRequired(getUserOidcMinimumVersion)"
-				:error-link="appLinks.user_oidc.installLink"
-				:error-link-label="messages.downloadAndEnableApp" />
+				:error-link-label="messages.installLatestVersionNow" />
 			<div class="authorization-settings--content">
 				<FieldValue v-if="isAuthorizationSettingsInViewMode"
 					is-required
@@ -52,18 +45,20 @@
 					</p>
 					<NcCheckboxRadioSwitch
 						:checked.sync="authorizationSetting.SSOProviderType"
-						:disabled="!isOIDCAppSupported || !hasEnabledSupportedOIDCApp"
+						:disabled="!hasEnabledSupportedUserOidcApp || !hasEnabledSupportedOIDCApp"
 						:value="SSO_PROVIDER_TYPE.nextcloudHub"
 						type="radio">
 						{{ messages.nextcloudHubProvider }}
 					</NcCheckboxRadioSwitch>
-					<ErrorLabel
-						v-if="!hasEnabledSupportedOIDCApp"
-						:error="`${messagesFmt.appNotEnabledOrSupported('oidc')}. ${messagesFmt.minimumVersionRequired(getMinSupportedOIDCVersion)}`"
-						:disabled="disableNCHubUnsupportedHint" />
+					<div class="error-container">
+						<ErrorLabel
+							v-if="!hasEnabledSupportedOIDCApp"
+							:error="messagesFmt.appNotEnabledOrUnsupported('oidc', getMinSupportedOIDCVersion)"
+							:disabled="disableNCHubUnsupportedHint" />
+					</div>
 					<NcCheckboxRadioSwitch
 						:checked.sync="authorizationSetting.SSOProviderType"
-						:disabled="!isOIDCAppInstalledAndEnabled || !isOIDCAppSupported"
+						:disabled="!hasEnabledSupportedUserOidcApp"
 						:value="SSO_PROVIDER_TYPE.external"
 						type="radio">
 						{{ messages.externalOIDCProvider }}
@@ -80,7 +75,7 @@
 					</p>
 					<NcSelect
 						input-id="provider-search-input"
-						:disabled="!isOIDCAppInstalledAndEnabled || !isOIDCAppSupported"
+						:disabled="!hasEnabledSupportedUserOidcApp"
 						:placeholder="t('integration_openproject', 'Select an OIDC provider')"
 						:options="registeredOidcProviders"
 						:value="getCurrentSelectedOIDCProvider"
@@ -122,7 +117,7 @@
 							v-model="authorizationSetting.currentTargetedAudienceClientIdSelected"
 							class="py-1"
 							is-required
-							:disabled="!isOIDCAppInstalledAndEnabled || !isOIDCAppSupported"
+							:disabled="!hasEnabledSupportedUserOidcApp"
 							:place-holder="messages.opClientId"
 							:label="messages.opClientId"
 							:hint-text="messages.opClientIdHintText" />
@@ -131,7 +126,7 @@
 			</div>
 			<div class="form-actions">
 				<NcButton v-if="isAuthorizationSettingsInViewMode"
-					:disabled="!isOIDCAppInstalledAndEnabled || !isOIDCAppSupported"
+					:disabled="!hasEnabledSupportedUserOidcApp"
 					data-test-id="reset-auth-settings-btn"
 					@click="setAuthorizationSettingInEditMode">
 					<template #icon>
@@ -793,11 +788,11 @@ export default {
 		getSSOProviderType() {
 			return this.authorizationSetting.SSOProviderType
 		},
-		isOIDCAppInstalledAndEnabled() {
-			return this.state.user_oidc_enabled
+		hasEnabledSupportedUserOidcApp() {
+			return this.state.apps.user_oidc.enabled && this.state.apps.user_oidc.supported
 		},
-		isOIDCAppSupported() {
-			return this.state.user_oidc_supported
+		getMinSupportedUserOidcVersion() {
+			return this.state.apps.user_oidc.minimum_version
 		},
 		hasEnabledSupportedOIDCApp() {
 			return this.state.apps.oidc.enabled && this.state.apps.oidc.supported
@@ -1513,10 +1508,6 @@ export default {
 			.radio-check {
 				font-weight: 500;
 			}
-			.oidc-app-check-description {
-				margin-left: 2.4rem;
-				font-size: 14px;
-			}
 		}
 	}
 	.authorization-settings {
@@ -1534,6 +1525,10 @@ export default {
 		.description {
 			margin-top: 0.1rem;
 		}
+	}
+	.error-container {
+		margin-left: 2.4rem;
+		font-size: 14px;
 	}
 }