+++ tests/jest/components/AdminSettings.spec.js
@@ -13,7 +13,7 @@ import flushPromises from 'flush-promises' // eslint-disable-line n/no-unpublish
 import AdminSettings from '../../../src/components/AdminSettings.vue'
 import { F_MODES, AUTH_METHOD, ADMIN_SETTINGS_FORM } from '../../../src/utils.js'
 import { appLinks } from '../../../src/constants/links.js'
-import { messages, messagesFmt } from '../../../src/constants/messages.js'
+import { messagesFmt } from '../../../src/constants/messages.js'
 
 jest.mock('@nextcloud/axios', () => {
 	const originalModule = jest.requireActual('@nextcloud/axios')
@@ -140,6 +140,11 @@ const appState = {
 			enabled: true,
 			supported: true,
 			minimum_version: '1.4.0',
+		},
+		user_oidc: {
+			enabled: true,
+			supported: true,
+			minimum_version: '2.0.0',
 		},
 	},
 }
@@ -579,6 +584,7 @@ describe('AdminSettings.vue', () => {
 				targeted_audience_client_id: null,
 				sso_provider_type: 'nextcloud_hub',
 			},
+			...appState,
 		}
 
 		describe('form complete: view mode', () => {
@@ -614,7 +620,7 @@ describe('AdminSettings.vue', () => {
 						authorization_settings: settings,
 					}
 					wrapper = getWrapper({
-						state: { ...state, ...authSettings, user_oidc_enabled: true, user_oidc_supported: true },
+						state: { ...state, ...authSettings },
 						form: {
 							serverHost: { complete: true },
 							authenticationMethod: { complete: true, value: AUTH_METHOD.OIDC },
@@ -642,7 +648,18 @@ describe('AdminSettings.vue', () => {
 			describe('unsupported user_oidc app enabled', () => {
 				beforeEach(async () => {
 					wrapper = getWrapper({
-						state: { ...state, ...authorizationSettingsState, user_oidc_enabled: true, user_oidc_supported: false },
+						state: {
+							...state,
+							...authorizationSettingsState,
+							apps: {
+								...appState.apps,
+								user_oidc: {
+									enabled: true,
+									supported: false,
+									minimum_version: appState.apps.user_oidc.minimum_version,
+								},
+							},
+						},
 						form: {
 							serverHost: { complete: true },
 							authenticationMethod: { complete: true, value: AUTH_METHOD.OIDC },
@@ -665,8 +682,7 @@ describe('AdminSettings.vue', () => {
 					expect(formHeader.exists()).toBe(true)
 					expect(formHeader.attributes().haserror).toBe('true')
 					expect(errorNote.exists()).toBe(true)
-					expect(errorNote.attributes().errortitle).toBe(messagesFmt.appNotSupported())
-					expect(errorNote.attributes().errormessage).toBe(messagesFmt.minimumVersionRequired())
+					expect(errorNote.attributes().errortitle).toBe(messagesFmt.appNotEnabledOrUnsupported())
 					expect(errorNote.attributes().errorlink).toBe(appLinks.user_oidc.installLink)
 				})
 			})
@@ -674,7 +690,18 @@ describe('AdminSettings.vue', () => {
 			describe('supported user_oidc app disabled', () => {
 				beforeEach(async () => {
 					wrapper = getWrapper({
-						state: { ...state, ...authorizationSettingsState, user_oidc_enabled: false, user_oidc_supported: true },
+						state: {
+							...state,
+							...authorizationSettingsState,
+							apps: {
+								...appState.apps,
+								user_oidc: {
+									enabled: false,
+									supported: true,
+									minimum_version: appState.apps.user_oidc.minimum_version,
+								},
+							},
+						},
 						form: {
 							serverHost: { complete: true },
 							authenticationMethod: { complete: true, value: AUTH_METHOD.OIDC },
@@ -697,8 +724,7 @@ describe('AdminSettings.vue', () => {
 					expect(formHeader.exists()).toBe(true)
 					expect(formHeader.attributes().haserror).toBe('true')
 					expect(errorNote.exists()).toBe(true)
-					expect(errorNote.attributes().errortitle).toBe(messagesFmt.appNotInstalled())
-					expect(errorNote.attributes().errormessage).toBe(messages.appRequiredForOIDCMethod)
+					expect(errorNote.attributes().errortitle).toBe(messagesFmt.appNotEnabledOrUnsupported())
 					expect(errorNote.attributes().errorlink).toBe(appLinks.user_oidc.installLink)
 				})
 			})
@@ -716,8 +742,6 @@ describe('AdminSettings.vue', () => {
 							state: {
 								...state,
 								...authSettings,
-								user_oidc_enabled: true,
-								user_oidc_supported: true,
 							},
 							form: {
 								serverHost: { complete: true },
@@ -756,8 +780,6 @@ describe('AdminSettings.vue', () => {
 							state: {
 								...state,
 								...authSettings,
-								user_oidc_enabled: true,
-								user_oidc_supported: true,
 							},
 							form: {
 								serverHost: { complete: true },
@@ -800,7 +822,10 @@ describe('AdminSettings.vue', () => {
 					authorization_settings: settings,
 				}
 				wrapper = getWrapper({
-					state: { ...state, ...authSettings, user_oidc_enabled: true, user_oidc_supported: true },
+					state: {
+						...state,
+						...authSettings,
+					},
 					form: {
 						serverHost: {
 							complete: true,
@@ -827,6 +852,7 @@ describe('AdminSettings.vue', () => {
 				jest.clearAllMocks()
 				wrapper = getMountedWrapper({
 					state: {
+						...appState,
 						openproject_instance_url: 'http://openproject.com',
 						authorization_method: AUTH_METHOD.OIDC,
 						authorization_settings: {
@@ -835,8 +861,6 @@ describe('AdminSettings.vue', () => {
 							targeted_audience_client_id: 'some-target-aud-client-id',
 							token_exchange: false,
 						},
-						user_oidc_enabled: true,
-						user_oidc_supported: true,
 					},
 					form: {
 						serverHost: { complete: true },
@@ -875,6 +899,7 @@ describe('AdminSettings.vue', () => {
 						wrapper = getMountedWrapper({
 							registeredOidcProviders: ['keycloak'],
 							state: {
+								...appState,
 								openproject_instance_url: 'http://openproject.com',
 								authorization_method: AUTH_METHOD.OIDC,
 								authorization_settings: {
@@ -882,8 +907,6 @@ describe('AdminSettings.vue', () => {
 									targeted_audience_client_id: 'some-target-aud-client-id',
 									sso_provider_type: 'external',
 								},
-								user_oidc_enabled: true,
-								user_oidc_supported: true,
 							},
 							form: {
 								serverHost: { complete: true },
@@ -926,6 +949,7 @@ describe('AdminSettings.vue', () => {
 						wrapper = getMountedWrapper({
 							registeredOidcProviders: ['keycloak'],
 							state: {
+								...appState,
 								openproject_instance_url: 'http://openproject.com',
 								authorization_method: AUTH_METHOD.OIDC,
 								authorization_settings: {
@@ -934,8 +958,6 @@ describe('AdminSettings.vue', () => {
 									sso_provider_type: 'external',
 									token_exchange: true,
 								},
-								user_oidc_enabled: true,
-								user_oidc_supported: true,
 							},
 							form: {
 								serverHost: { complete: true },
@@ -1018,9 +1040,8 @@ describe('AdminSettings.vue', () => {
 								targeted_audience_client_id: 'some-target-aud-client-id',
 								sso_provider_type: 'nextcloud_hub',
 							},
-							user_oidc_enabled: true,
-							user_oidc_supported: true,
 							apps: {
+								...appState.apps,
 								oidc: {
 									enabled: false,
 									supported: false,
@@ -1050,7 +1071,7 @@ describe('AdminSettings.vue', () => {
 					expect(ncProviderRadio.attributes().disabled).toBe('true')
 					expect(ncProviderRadio.attributes().checked).toBe('nextcloud_hub')
 					expect(externalProviderRadio.attributes().disabled).toBe(undefined)
-					expect(errorLabel.attributes().error).toBe(`${messagesFmt.appNotEnabledOrSupported('oidc')}. ${messagesFmt.minimumVersionRequired('oidc')}`)
+					expect(errorLabel.attributes().error).toBe(messagesFmt.appNotEnabledOrUnsupported('oidc'))
 					expect(errorLabel.attributes().disabled).toBe(undefined)
 				})
 			})
@@ -1062,7 +1083,10 @@ describe('AdminSettings.vue', () => {
 			describe('Supported user_oidc app enabled', () => {
 				beforeEach(async () => {
 					wrapper = getWrapper({
-						state: { ...state, user_oidc_enabled: true, user_oidc_supported: true },
+						state: {
+							...state,
+							...appState,
+						},
 						form: {
 							serverHost: { complete: true },
 							authenticationMethod: { complete: true, value: AUTH_METHOD.OIDC },
@@ -1098,6 +1122,7 @@ describe('AdminSettings.vue', () => {
 				it('should disable "save" button for empty "targeted_audience_client_id"', () => {
 					const wrapper = getWrapper({
 						state: {
+							...appState,
 							openproject_instance_url: 'http://openproject.com',
 							authorization_method: AUTH_METHOD.OIDC,
 							authorization_settings: {
@@ -1117,14 +1142,13 @@ describe('AdminSettings.vue', () => {
 				describe('external SSO provider', () => {
 					const wrapper = getWrapper({
 						state: {
+							...appState,
 							openproject_instance_url: 'http://openproject.com',
 							authorization_method: AUTH_METHOD.OIDC,
 							authorization_settings: {
 								oidc_provider: '',
 								targeted_audience_client_id: '',
 							},
-							user_oidc_enabled: true,
-							user_oidc_supported: true,
 						},
 						authorizationSetting: {
 							SSOProviderType: 'external',
@@ -1156,14 +1180,13 @@ describe('AdminSettings.vue', () => {
 						const wrapper = getMountedWrapper({
 							registeredOidcProviders: ['keycloak'],
 							state: {
+								...appState,
 								openproject_instance_url: 'http://openproject.com',
 								authorization_method: AUTH_METHOD.OIDC,
 								authorization_settings: {
 									oidc_provider: '',
 									targeted_audience_client_id: '',
 								},
-								user_oidc_enabled: true,
-								user_oidc_supported: true,
 							},
 							authorizationSetting: {
 								SSOProviderType: 'external',
@@ -1188,14 +1211,13 @@ describe('AdminSettings.vue', () => {
 							wrapper = getMountedWrapper({
 								registeredOidcProviders: ['keycloak'],
 								state: {
+									...appState,
 									openproject_instance_url: 'http://openproject.com',
 									authorization_method: AUTH_METHOD.OIDC,
 									authorization_settings: {
 										oidc_provider: '',
 										targeted_audience_client_id: '',
 									},
-									user_oidc_enabled: true,
-									user_oidc_supported: true,
 								},
 								authorizationSetting: {
 									SSOProviderType: 'external',
@@ -1236,8 +1258,7 @@ describe('AdminSettings.vue', () => {
 							wrapper = getMountedWrapper({
 								state: {
 									...state,
-									user_oidc_enabled: true,
-									user_oidc_supported: true,
+									...appState,
 								},
 								form: {
 									serverHost: { complete: true },
@@ -1288,8 +1309,7 @@ describe('AdminSettings.vue', () => {
 								registeredOidcProviders: ['keycloak'],
 								state: {
 									...state,
-									user_oidc_enabled: true,
-									user_oidc_supported: true,
+									...appState,
 									authorization_settings: {},
 								},
 							 })
@@ -1352,8 +1372,14 @@ describe('AdminSettings.vue', () => {
 								oidc_provider: '',
 								targeted_audience_client_id: '',
 							},
-							user_oidc_enabled: false,
-							user_oidc_supported: true,
+							apps: {
+								...appState.apps,
+								user_oidc: {
+									enabled: false,
+									supported: true,
+									minimum_version: appState.apps.user_oidc.minimum_version,
+								},
+							},
 						},
 						form: {
 							serverHost: { complete: true },
@@ -1369,8 +1395,7 @@ describe('AdminSettings.vue', () => {
 					expect(formHeaderError.exists()).toBe(true)
 					expect(formHeaderError.attributes().haserror).toBe('true')
 					expect(errorNote.exists()).toBe(true)
-					expect(errorNote.attributes().errortitle).toBe(messagesFmt.appNotInstalled())
-					expect(errorNote.attributes().errormessage).toBe(messages.appRequiredForOIDCMethod)
+					expect(errorNote.attributes().errortitle).toBe(messagesFmt.appNotEnabledOrUnsupported())
 					expect(errorNote.attributes().errorlink).toBe(appLinks.user_oidc.installLink)
 				})
 				it('should disable form elements', () => {
@@ -1395,8 +1420,14 @@ describe('AdminSettings.vue', () => {
 								oidc_provider: '',
 								targeted_audience_client_id: '',
 							},
-							user_oidc_enabled: true,
-							user_oidc_supported: false,
+							apps: {
+								...appState.apps,
+								user_oidc: {
+									enabled: true,
+									supported: false,
+									minimum_version: appState.apps.user_oidc.minimum_version,
+								},
+							},
 						},
 						form: {
 							serverHost: { complete: true },
@@ -1412,8 +1443,7 @@ describe('AdminSettings.vue', () => {
 					expect(formHeaderError.exists()).toBe(true)
 					expect(formHeaderError.attributes().haserror).toBe('true')
 					expect(errorNote.exists()).toBe(true)
-					expect(errorNote.attributes().errortitle).toBe(messagesFmt.appNotSupported())
-					expect(errorNote.attributes().errormessage).toBe(messagesFmt.minimumVersionRequired())
+					expect(errorNote.attributes().errortitle).toBe(messagesFmt.appNotEnabledOrUnsupported())
 					expect(errorNote.attributes().errorlink).toBe(appLinks.user_oidc.installLink)
 				})
 				it('should disable form elements', () => {
@@ -1438,9 +1468,8 @@ describe('AdminSettings.vue', () => {
 								oidc_provider: '',
 								targeted_audience_client_id: '',
 							},
-							user_oidc_enabled: true,
-							user_oidc_supported: true,
 							apps: {
+								...appState.apps,
 								oidc: {
 									enabled: false,
 									supported: false,
@@ -1463,7 +1492,7 @@ describe('AdminSettings.vue', () => {
 					expect(ncProviderRadio.attributes().disabled).toBe('true')
 					expect(ncProviderRadio.attributes().checked).toBe('external')
 					expect(externalProviderRadio.attributes().disabled).toBe(undefined)
-					expect(errorLabel.attributes().error).toBe(`${messagesFmt.appNotEnabledOrSupported('oidc')}. ${messagesFmt.minimumVersionRequired('oidc')}`)
+					expect(errorLabel.attributes().error).toBe(messagesFmt.appNotEnabledOrUnsupported('oidc'))
 					expect(errorLabel.attributes().disabled).toBe('true')
 				})
 			})